+++ docs/systemverilog_parser_book-html/searcher-c2a407aa.js
@@ -437,7 +437,7 @@ window.search = window.search || {};
         if (yes) {
             loadSearchScript(
                 window.path_to_searchindex_js ||
-                path_to_root + 'searchindex-f2121879.js',
+                path_to_root + 'searchindex-bbdd65ba.js',
                 'mdbook-search-index');
             search_wrap.classList.remove('hidden');
             searchicon.setAttribute('aria-expanded', 'true');